// newcode/libs/modules/estimate/main/src/lib/model/enums/estimate-main-resource-filter-type.enum.ts
/*
 * Copyright(c) RIB Software GmbH
 */
export enum EstMainResFilterType {
	AllResources = 'allResources',
	ResWithRule = 'resWithRule',
	ResWithoutRule = 'resWithoutRule',
	ResWithoutDisable = 'resWithoutDisable',
}

export enum EstMainResFilterTitleType {
	AllResources = 'estimate.main.showAllResources',
	ResWithRule = 'estimate.main.showResourcesWithRule',
	ResWithoutRule = 'estimate.main.showResourcesWithoutRule',
	ResWithoutDisable = 'estimate.main.showResourcesWithoutDisable',
}